requirements.txt
numpy>=1.19.4
matplotlib>=3.3.3
gensim>=3.8.3
py-ecg-detectors>=1.0.2
sklearn>=0.0
